requirements.txt
FastAPI == 0.79.1
gunicorn == 20.1.0
uvicorn == 0.18.2
BigQuery-Python == 1.15.0
google-cloud-bigquery == 3.3.2
google-cloud-storage == 2.5.0